// frontend/src/sections/items/itemPreferences.generated.ts
/* eslint-disable */
// THIS FILE IS GENERATED by codegen/plugin.js — do not edit by hand.
import type { TypedDocument } from "../../api/graphql";

export type ItemPreferencesVariables = {
  storeId: string;
};

export type ItemPreferencesResult = {
  preferences: {
  manageVaccinesInDoses: boolean;
  adjustForNumberOfDaysOutOfStock: boolean;
  numberOfMonthsToCheckForConsumptionWhenCalculatingOutOfStockProducts: number;
  numberOfMonthsThresholdToShowLowStockAlertsForProducts: number;
};
};

export const ItemPreferences = {
  query: "query itemPreferences($storeId: String!) {\n  preferences(storeId: $storeId) {\n    manageVaccinesInDoses\n    adjustForNumberOfDaysOutOfStock\n    numberOfMonthsToCheckForConsumptionWhenCalculatingOutOfStockProducts\n    numberOfMonthsThresholdToShowLowStockAlertsForProducts\n  }\n}",
} as TypedDocument<ItemPreferencesResult, ItemPreferencesVariables>;