// frontend/src/sections/items/list/itemMasterLists.generated.ts
/* eslint-disable */
// THIS FILE IS GENERATED by codegen/plugin.js — do not edit by hand.
import type { TypedDocument } from "../../../api/graphql";

export type ItemMasterListsVariables = {
  storeId: string;
};

export type ItemMasterListsResult = {
  masterLists: ({
  __typename: "MasterListConnector";
} & {
  totalCount: number;
  nodes: Array<{
  id: string;
  name: string;
}>;
});
};

export const ItemMasterLists = {
  query: "query itemMasterLists($storeId: String!) {\n  masterLists(storeId: $storeId, filter: {existsForStoreId: {equalTo: $storeId}}) {\n    ... on MasterListConnector {\n      __typename\n      totalCount\n      nodes {\n        id\n        name\n      }\n    }\n  }\n}",
} as TypedDocument<ItemMasterListsResult, ItemMasterListsVariables>;